requirements.txt
ultralytics>=8.2.0
opencv-python>=4.10.0
numpy>=1.26.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.37.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.8.0
PyYAML>=6.0.2
